# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: clinics
    description: Clinic information and patient-facing availability
  - name: appointments
    description: Patient appointment booking and reminders
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /clinic:
    get:
      operationId: getClinic
      tags: [clinics]
      summary: Get clinic information
      responses:
        "200":
          description: Clinic information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinic"
  /doctors:
    get:
      operationId: listDoctors
      tags: [clinics]
      summary: List clinic doctors
      responses:
        "200":
          description: Doctors available for appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Doctor"
  /appointment-slots:
    get:
      operationId: listAppointmentSlots
      tags: [clinics]
      summary: List available appointment slots
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: doctorId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Available and booked appointment slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppointmentSlot"
  /appointments:
    get:
      operationId: listPatientAppointments
      tags: [appointments]
      summary: List appointments for a patient
      parameters:
        - name: phone
          in: query
          required: true
          schema:
            type: string
            minLength: 7
      responses:
        "200":
          description: Patient appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Appointment booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "409":
          description: Slot is no longer available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get an appointment confirmation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment confirmation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /appointments/{id}/reminder:
    patch:
      operationId: updateAppointmentReminder
      tags: [appointments]
      summary: Update appointment reminder preference
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderInput"
      responses:
        "200":
          description: Appointment with updated reminder preference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /clinic-summary:
    get:
      operationId: getClinicSummary
      tags: [clinics]
      summary: Get a patient-facing clinic summary
      responses:
        "200":
          description: Clinic summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicSummary"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Clinic:
      type: object
      required: [id, name, tagline, address, phone, openingHours, services]
      properties:
        id:
          type: integer
        name:
          type: string
        tagline:
          type: string
        address:
          type: string
        phone:
          type: string
        openingHours:
          type: string
        services:
          type: array
          items:
            type: string
    Doctor:
      type: object
      required: [id, name, specialty, initials, experience]
      properties:
        id:
          type: integer
        name:
          type: string
        specialty:
          type: string
        initials:
          type: string
        experience:
          type: string
    AppointmentSlot:
      type: object
      required: [id, doctorId, doctorName, specialty, date, startTime, endTime, status]
      properties:
        id:
          type: integer
        doctorId:
          type: integer
        doctorName:
          type: string
        specialty:
          type: string
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        status:
          type: string
          enum: [available, booked]
    Appointment:
      type: object
      required: [id, confirmationCode, patientName, patientPhone, doctorId, doctorName, specialty, date, startTime, endTime, status, reminderEnabled, createdAt]
      properties:
        id:
          type: integer
        confirmationCode:
          type: string
        patientName:
          type: string
        patientPhone:
          type: string
        patientEmail:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
        doctorId:
          type: integer
        doctorName:
          type: string
        specialty:
          type: string
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        status:
          type: string
          enum: [confirmed, completed, cancelled]
        reminderEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AppointmentInput:
      type: object
      required: [slotId, patientName, patientPhone]
      properties:
        slotId:
          type: integer
        patientName:
          type: string
          minLength: 2
        patientPhone:
          type: string
          minLength: 7
        patientEmail:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
        reminderEnabled:
          type: boolean
          default: true
    ReminderInput:
      type: object
      required: [reminderEnabled]
      properties:
        reminderEnabled:
          type: boolean
    ClinicSummary:
      type: object
      required: [nextAvailableDate, availableToday, totalDoctors, services]
      properties:
        nextAvailableDate:
          type: string
          format: date
        availableToday:
          type: integer
        totalDoctors:
          type: integer
        services:
          type: array
          items:
            type: string
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string