# lib/api-zod/src/generated/types/appointment.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentStatus } from './appointmentStatus';

export interface Appointment {
  id: number;
  confirmationCode: string;
  patientName: string;
  patientPhone: string;
  /** @nullable */
  patientEmail?: string | null;
  /** @nullable */
  reason?: string | null;
  doctorId: number;
  doctorName: string;
  specialty: string;
  date: Date;
  startTime: string;
  endTime: string;
  status: AppointmentStatus;
  reminderEnabled: boolean;
  createdAt: Date;
}