# lib/api-zod/src/generated/types/appointmentInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AppointmentInput {
  slotId: number;
  /** @minLength 2 */
  patientName: string;
  /** @minLength 7 */
  patientPhone: string;
  /** @nullable */
  patientEmail?: string | null;
  /** @nullable */
  reason?: string | null;
  reminderEnabled?: boolean;
}