# lib/api-zod/src/generated/types/appointmentSlot.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentSlotStatus } from './appointmentSlotStatus';

export interface AppointmentSlot {
  id: number;
  doctorId: number;
  doctorName: string;
  specialty: string;
  date: Date;
  startTime: string;
  endTime: string;
  status: AppointmentSlotStatus;
}