# lib/api-zod/src/generated/types/clinic.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Clinic {
  id: number;
  name: string;
  tagline: string;
  address: string;
  phone: string;
  openingHours: string;
  services: string[];
}