# lib/api-zod/src/generated/types/clinicSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ClinicSummary {
  nextAvailableDate: Date;
  availableToday: number;
  totalDoctors: number;
  services: string[];
}