# lib/api-zod/src/generated/types/doctor.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Doctor {
  id: number;
  name: string;
  specialty: string;
  initials: string;
  experience: string;
}